Move frame dumper to core
